% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{load_and_validate_data}
\alias{load_and_validate_data}
\title{Load, Validate, and Clip Top Soil Data}
\usage{
load_and_validate_data(
  file_path = system.file("extdata", "soil_data/LUCAS-SOIL-2018.csv", package =
    "SoilGeoStats"),
  x_col = "TH_LONG",
  y_col = "TH_LAT",
  target_col = "OC",
  covariate_cols = c("N", "Elev"),
  shape_path = system.file("extdata", "shape_file/Bav_Boundary.shp", package =
    "SoilGeoStats")
)
}
\arguments{
\item{file_path}{Character string. Relative path to the soil CSV file.}

\item{x_col}{Character string. X coordinate column.}

\item{y_col}{Character string. Y coordinate column.}

\item{target_col}{Character string. The primary target variable ("OC").}

\item{covariate_cols}{Character vector. Names of covariate columns.}

\item{shape_path}{Character string. Relative path to the boundary shapefile.}
}
\value{
An 'sf' spatial data frame object with validated numeric columns, clipped to the shapefile if available.
}
\description{
Load, Validate, and Clip Top Soil Data
}
