% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_spatial_grid}
\alias{new_spatial_grid}
\title{Create a Spatial Grid Object Constructor}
\usage{
new_spatial_grid(
  grid_data,
  method,
  target_variable,
  original_points,
  covariate_cols = NULL
)
}
\arguments{
\item{grid_data}{An 'sf' spatial object containing interpolated prediction rows.}

\item{method}{Character string. The model engine used ("IDW" or "Regression Kriging").}

\item{target_variable}{Character string. The column modeled ("OC").}

\item{original_points}{The original 'sf' sample point dataset used to build the model.}

\item{covariate_cols}{Character vector. Covariates used if running Regression Kriging.}
}
\value{
A S3 object of class 'spatial_grid'.
}
\description{
Create a Spatial Grid Object Constructor
}
