% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.spatial_grid}
\alias{plot.spatial_grid}
\title{Plot Method for Spatial Grid Objects}
\usage{
\method{plot}{spatial_grid}(x, boundary_data, ...)
}
\arguments{
\item{x}{An object of class 'spatial_grid'.}

\item{boundary_data}{The 'sf' spatial object used as the clipping mask (Bavaria boundary).}

\item{...}{Unused additional parameters.}
}
\description{
Plot Method for Spatial Grid Objects
}
