% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{calculate_spatial_distances}
\alias{calculate_spatial_distances}
\title{Evaluate Spatial Sample Point Grid Characteristics}
\usage{
calculate_spatial_distances(spatial_data, target_col)
}
\arguments{
\item{spatial_data}{The 'sf' object generated from load_and_validate_data().}

\item{target_col}{Character string. The name of the soil variable column of interest (OC for this context).}
}
\value{
A named list containing structural summaries of geographic distances and result of exploratory data analysis.
}
\description{
Evaluate Spatial Sample Point Grid Characteristics
}
